refactor: ADMIN DIR 위치 수정
